Added event manager
Added the eventManager object to monitor the number of minions defeated.
After the player defeats 3 of the regular minions, and after they defeat the super minion, a "Level Up" particle effect will appear above the player's head.
Tweaked the animator graph so that the animations don't have to wait to finish before transitioning to the next state.
Adjusted the collider box sizes and positions to make it easier to hit.
Adjusted the position of the overlapSphere to be further out from the player.
Added the skybox.