Added Defending poses
Added a defending motion and controls to the player character. The player can only defend while they are on the ground. No other functionality is implemented yet. The idea is that the shield will have a collision box that the enemies will treat like a solid object, and prevent them from getting closer, as well as block projectiles.